+++ SubEthaEdit-Mac/Source/NSStringSEEAdditions.h
@@ -63,9 +63,13 @@ typedef enum _LineEnding {
 - (NSMutableString *)XHTMLStringWithAttributeMapping:(NSDictionary *)anAttributeMapping forUTF8:(BOOL)forUTF8;
 - (NSRange)TCM_fullLengthRange;
 - (NSUInteger)TCM_positionOfMatchingBracketToPosition:(NSUInteger)position bracketSettings:(TCMBracketSettings *)aBracketSettings;
+
++ (NSDictionary *)SEE_lineNumberAttributesWithFontSize:(CGFloat)fontSize baseAttributes:(NSDictionary *)baseAttribtues;
++ (BOOL)SEE_usesModernSystemFont;
 @end
 
 
 @interface NSFont (NSFontSEEAdditions)
++ (NSFont *)SEE_lineNumbersFontOfSize:(CGFloat)size;
 - (NSFont *)SEE_fontByAddingMonoSpaceNumbersFeature;
 @end